EHV-53 - Available hours is updated realtime
Conflicts:
	eHour-common/src/main/java/net/rrm/ehour/domain/TimesheetEntry.java
	eHour-service/src/test/java/net/rrm/ehour/timesheet/service/TimesheetServiceTest.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/common/FormHighlighter.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/dto/Timesheet.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/dto/TimesheetRow.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/panel/TimesheetPanel.html
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/panel/TimesheetPanel.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/panel/TimesheetRowList.java
	eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/util/TimesheetAssembler.java

+++ eHour-wicketweb/src/main/java/net/rrm/ehour/ui/timesheet/dto/ProjectTotalModel.java
@@ -32,19 +32,19 @@ public class ProjectTotalModel extends Model<Float>
 	public Float getObject()
 	{
 		float	totalHours = 0;
-		
-		for (int i = 0; 
+
+		for (int i = 0;
 				i < row.getTimesheetCells().length;
 				i++)
 		{
-			if (row.getTimesheetCells()[i] != null 
-					&& row.getTimesheetCells()[i].getTimesheetEntry() != null 
+			if (row.getTimesheetCells()[i] != null
+					&& row.getTimesheetCells()[i].getTimesheetEntry() != null
 					&& row.getTimesheetCells()[i].getTimesheetEntry().getHours() != null)
 			{
-				totalHours += row.getTimesheetCells()[i].getTimesheetEntry().getHours().floatValue();
-			}			
+				totalHours += row.getTimesheetCells()[i].getTimesheetEntry().getHours();
+			}
 		}
-		
-		return Float.valueOf(totalHours);
+
+		return totalHours;
 	}
 }
